Solved AoC2020 Day 24 Part 1

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day24/Day24Test.kt
@@ -10,7 +10,7 @@ class Day24Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         LobbyLayout(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
@@ -22,7 +22,7 @@ class Day24Test {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day24/day24part1sample.txt", 10L)
+                Arguments.of("de/havox_design/aoc2020/day24/day24part1sample.txt", 10)
             )
 
         @JvmStatic
